Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I9205f5f3183eff8059e7589b6a3a2bbf36f8b60a
Partial-Bug: #1764470
(cherry picked from commit 1f606ff46173bd04990891b2d833b1443d6e04a7)