Set container_name as host_var in test inventory
The connection plugin no longer falls back on using the
inventory_hostname as the container_name. Set container_name as a host
var for each container in the test inventory.

In this role's case there are no containers, so I've just removed
the group_var so that the test setup is uniform with other roles
which do use containers.

Change-Id: Iab5c434cd40b909b8b43f52cd9f50cae83c1c715